api/requirements.txt
scikit-learn===1.2.2
selenium
webdriver_manager
fastapi
pickle5
uvicorn
python-dotenv